requirements.txt
Flask~=3.1.2
Werkzeug~=3.1.4
pillow~=12.0.0